新增“退出堆栈式上下文管理器示例”
